// front/src/api/types.ts
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};
// ...
export type LivreInput = {
  design: Scalars['String'];
  auteur: Scalars['String'];
  dateEdition: Scalars['String'];
  disponible: Scalars['Boolean'];
};
// ...
export type MutationAddLivreArgs = {
  input: LivreInput;
};